refactor(data): Alinear los modelos de datos y los tipos de ID para la sincronización remota
Esta confirmación refactoriza la capa de datos para unificar los tipos de ID, mejorar las operaciones de la API y fortalecer la lógica de sincronización sin conexión.

### Cambios en los Modelos de Datos y Entidades
- **Tipos de ID unificados:** Los tipos de ID en `UserCar`, `MaintenanceTask` y `MaintenanceHistory` y sus correspondientes `Entity` han sido cambiados de `String` o `Long` a `Int` para alinearse con los IDs numéricos de la API remota.
- **Entidades de Base de Datos:** Las claves primarias (`id`) para `MaintenanceTaskEntity` y `MaintenanceHistoryEntity` ahora son `Int` con autogeneración, desacoplando el ID local del `remoteId`.
- **Mapeadores de DTO y Dominio:** Los mapeadores (`DomainDtoMappers.kt`) han sido actualizados para manejar la conversión entre los nuevos IDs `Int` en las entidades/DTO y los IDs correspondientes en los modelos de dominio.

### Mejoras en DAO y Sincronización
- **Nuevos Métodos DAO:**
  - `MaintenanceTaskDao` y `MaintenanceHistoryDao` ahora incluyen los métodos `upsertAll()` y `clearForCar(carId)`. Estos permiten el refresco por lotes de datos locales durante la sincronización remota.
- **Lógica de Sincronización:**
  - `MaintenanceRepositoryImpl` implementa `syncFromRemote`, que ahora obtiene tareas e historiales de la API, limpia los datos locales para el coche (`clearForCar`), y luego inserta los nuevos datos (`upsertAll`).
  - La lógica de `pushPending` ha sido implementada para sincronizar creaciones, actualizaciones y eliminaciones locales con el servidor remoto.

### Refactorización de la Capa de Red
- **API Services:**
  - `CarApiService` ahora utiliza el `UserCarDto` directamente como el cuerpo de la solicitud para las operaciones de creación y actualización (`createCar`, `updateCar`), eliminando los `CreateUserCarRequest` y `UpdateUserCarRequest` que eran redundantes.
- **Data Sources:**
  - Todos los `RemoteDataSource` (`Car`, `Maintenance`, `VehicleCatalog`) han sido refactorizados para usar tipos explícitos en las respuestas de la API y en los mapeos de listas, mejorando la seguridad de tipos y la claridad.
  - Se añadió lógica de filtrado en `VehicleCatalogRemoteDataSource` para asegurar que los modelos y años devueltos coincidan con el `brandId` y `modelId` solicitados.

+++ app/src/main/java/edu/ucne/loginapi/data/entity/UserCarEntity.kt
@@ -14,6 +14,6 @@ data class UserCarEntity(
     val fuelType: String,
     val usageType: String,
     val isCurrent: Boolean,
-    val remoteId: Long? = null,
+    val remoteId: Int? = null,
     val pendingSync: Boolean = false
 )
